pyproject.toml
[project]
name = "golduck"
version = "0.1.0"
description = "A Python API for querying Delta tables using DuckDB with OpenTelemetry monitoring."
authors = [
    { name = "Achraf Cherki Elidrissi", email = "achraf.c.e.i@gmail.com" }
]
license = "MIT"
requires-python = ">=3.12"


dependencies = [
    "fastapi",
    "uvicorn",
    "duckdb",
    "polars",
    "pytz",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-logging",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-exporter-otlp",
    "pydantic",
    "pydantic-settings",
    "pyarrow",
    "psutil",
    "redis>=4.6.0",
]

[project.scripts]
golduck = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
